lib/backend/python_backend/requirements.txt
# Mangrove Watch AI Backend Requirements
# Python 3.13+ Compatible
# Last Updated: 2025-01-30

# ===== Core Framework =====
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ===== Database =====
motor==3.6.0
pymongo==4.9.1

# ===== Authentication & Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1

# ===== Google Services =====
google-generativeai==0.8.3
earthengine-api==1.1.2
google-api-python-client==2.108.0
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# ===== Image Processing =====
Pillow==10.4.0

# ===== Machine Learning =====
scikit-learn==1.5.2
joblib==1.3.2
numpy==2.1.1
pandas==2.2.3
# Note: TensorFlow may need separate installation for Python 3.13
# tensorflow==2.15.0  # Install separately if needed: pip install tensorflow

# ===== API Models & Validation =====
pydantic==2.10.2
pydantic[email]==2.10.2
pydantic-settings==2.7.0

# ===== Data Visualization (Optional) =====
matplotlib==3.9.2
# pyrsgis==0.4.1  # May have compatibility issues with Python 3.13

# ===== HTTP & Async Operations =====
httpx==0.27.2
aiofiles==24.1.0
starlette==0.41.3

# ===== Utilities =====
python-dateutil==2.8.2
typing-extensions==4.12.2

# ===== Additional Dependencies =====
# Required by various packages
cryptography==43.0.3
bcrypt==4.2.1
click==8.1.7
h11==0.14.0
# httptools==0.6.1  # Optional - may have issues on Windows
python-json-logger==2.0.7
six==1.16.0
# watchfiles==1.0.3  # Optional - may have issues on Windows
websockets==14.1
pyyaml==6.0.2

# ===== Development Tools (Optional) =====
# Uncomment if needed for development
# pytest==8.3.4
# pytest-asyncio==0.25.2
# black==24.10.0
# flake8==7.1.1
# mypy==1.14.1
# ipython==8.31.0

# ===== Production Server (Optional) =====
# Uncomment for production deployment
# gunicorn==23.0.0
# redis==5.2.1
# celery==5.4.0

# ===== Monitoring (Optional) =====
# Uncomment for production monitoring
# prometheus-client==0.21.0
# sentry-sdk==2.19.0
